+++ equalsverifier-core/src/main/java/nl/jqno/equalsverifier/internal/reflection/RecordProbe.java
@@ -36,16 +36,15 @@ public class RecordProbe<T> {
 
     private String buildMessage(Throwable e, List<?> params) {
         String msg = "Record: failed to run constructor for record type " + type.getName()
-                + "\n   These were the values passed to the constructor: " + params;
+                + "\nThese were the values passed to the constructor: " + params;
 
         if (e.getCause() instanceof NullPointerException) {
-            return msg + "\n   If the record does not accept null values for its constructor parameters,"
+            return msg + "\nIf the record does not accept null values for its constructor parameters,"
                     + " consider suppressing Warning.NULL_FIELDS.";
         }
 
-        msg += """
-               \n   If the record does not accept any of the given values for its constructor parameters, \
-               consider providing prefab values for the types of those fields.""";
+        msg += "\nIf the record does not accept any of the given values for its constructor parameters,"
+                + " consider providing prefab values for the types of those fields.";
         return msg;
     }
 }